+++ Runtime/IGU/IGUObjects/IGULayouts/IGUVerticalLayout.cs
@@ -2,14 +2,13 @@
 using UnityEngine;
 using Cobilas.Collections;
 using Cobilas.Unity.Graphics.IGU.Elements;
+using Cobilas.Unity.Graphics.IGU.Interfaces;
 
 namespace Cobilas.Unity.Graphics.IGU.Layouts {
-    public sealed class IGUVerticalLayout : IGULayout, ISerializationCallbackReceiver {
+    public sealed class IGUVerticalLayout : IGULayout, IIGUSerializationCallbackReceiver {
 
         [SerializeField]
         private CellIGUObject[] objects;
-        [SerializeField]
-        private bool AfterDeserialize = false;
         private VerticalLayoutCellCursor cursor;
         private event Action<CellCursor> sub_OnIGU;
 
@@ -30,12 +29,6 @@ namespace Cobilas.Unity.Graphics.IGU.Layouts {
             CellSize = Vector2.one * 100f;
             myRect = IGURect.DefaultButton;
             myColor = IGUColor.DefaultBoxColor;
-        }
-
-        protected override void OnEnable() {
-            if (!AfterDeserialize) return;
-            AfterDeserialize = false;
-            RefreshOnIGUEvent();
         }
 
         public override void OnIGU() {
@@ -97,10 +90,11 @@ namespace Cobilas.Unity.Graphics.IGU.Layouts {
                 sub_OnIGU += objects[I].OnIGU;
         }
 
-        void ISerializationCallbackReceiver.OnAfterDeserialize()
-            => AfterDeserialize = true;
-
-        void ISerializationCallbackReceiver.OnBeforeSerialize() { }
+        void IIGUSerializationCallbackReceiver.Reserialization() {
+#if UNITY_EDITOR
+            RefreshOnIGUEvent();
+#endif
+        }
 
         [Serializable]
         private sealed class VerticalLayoutCellCursor : CellCursor {