+++ Runtime/IGU/IGUObjects/IGUObjectBase/IGUTextObject.cs
@@ -2,9 +2,9 @@
 
 namespace Cobilas.Unity.Graphics.IGU.Elements {
     public abstract class IGUTextObject : IGUObject {
+        protected GUIStyle tooltipStyle;
         [SerializeField] protected bool useTooltip;
         [SerializeField] protected IGUContent content;
-        /*[SerializeField]*/ protected GUIStyle tooltipStyle;
 
         private static readonly GUIContent GUIContentTemp = new GUIContent();
 
@@ -15,8 +15,13 @@ namespace Cobilas.Unity.Graphics.IGU.Elements {
         public string ToolTip { get => content.Tooltip; set => content.Tooltip = value; }
         public GUIStyle TooltipStyle { get => tooltipStyle; set => tooltipStyle = value; }
 
+        protected override void Awake() {
+            base.Awake();
+            useTooltip = false;
+            content = new IGUContent();
+        }
+
         public override void OnIGU() => base.OnIGU();
-        protected override void Awake() => base.Awake();
         protected override void OnEnable() => base.OnEnable();
         protected override void OnDisable() => base.OnDisable();
         protected override void OnIGUDestroy() => base.OnIGUDestroy();