+++ Runtime/IGU/Interfaces/IIGUSerializationCallbackReceiver.cs
@@ -1,0 +1,5 @@
+﻿namespace Cobilas.Unity.Graphics.IGU.Interfaces {
+    public interface IIGUSerializationCallbackReceiver {
+        void Reserialization();
+    }
+}